+++ conftest.py
@@ -1,8 +1,8 @@
 import pytest
 
-from mysql2ch.factory import init
+from synch.factory import init
 
 
 @pytest.fixture(scope="session", autouse=True)
 def initialize_tests():
-    init('./mysql2ch.ini')
+    init('synch.ini')